requirements.txt
langchain-anthropic
langchain-ollama
langgraph
python-dotenv
pathlib